requirements.txt
alembic==1.12.0
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask===2.3.1
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
PyJWT==2.8.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.0
WTForms==3.1.0